services/news-signal/pyproject.toml
[project]
name = "news-signal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core>=0.12.5",
    "llama-index-llms-anthropic>=0.6.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.4.0",
    "quixstreams>=3.4.0",
    "loguru>=0.7.3",
    "llama-index-llms-ollama>=0.5.0",
    "fire>=0.7.0",
]

[dependency-groups]
gpu-instance= [
    "unsloth>=2024.12.4",
    "datasets>=3.2.0",
    "comet-ml>=3.47.5",
    "setuptools>=75.6.0",
]
